Move oppfolgingsplanlsp folder out of personoppgave
