+++ src/test/kotlin/no/nav/syfo/kafka/KafkaITSpek.kt
@@ -1,7 +1,7 @@
 package no.nav.syfo.kafka
 
-import no.nav.syfo.personoppgave.oppfolgingsplanlps.kafka.KOppfolgingsplanLPS
-import no.nav.syfo.personoppgave.oppfolgingsplanlps.kafka.OPPFOLGINGSPLAN_LPS_NAV_TOPIC
+import no.nav.syfo.oppfolgingsplanlps.kafka.KOppfolgingsplanLPS
+import no.nav.syfo.oppfolgingsplanlps.kafka.OPPFOLGINGSPLAN_LPS_NAV_TOPIC
 import no.nav.syfo.testutil.*
 import org.amshove.kluent.shouldBeEqualTo
 import org.spekframework.spek2.Spek
