+++ src/test/kotlin/no/nav/syfo/testutil/KOppfolgingsplanLPSNAVGenerator.kt
@@ -1,6 +1,6 @@
 package no.nav.syfo.testutil
 
-import no.nav.syfo.personoppgave.oppfolgingsplanlps.kafka.KOppfolgingsplanLPS
+import no.nav.syfo.oppfolgingsplanlps.kafka.KOppfolgingsplanLPS
 import no.nav.syfo.testutil.UserConstants.ARBEIDSTAKER_FNR
 import no.nav.syfo.testutil.UserConstants.VIRKSOMHETSNUMMER
 import java.time.LocalDate
